+++ HISTORY.rst
@@ -2,6 +2,12 @@
 History
 =======
 
+0.1.2 (2020-11-29)
+------------------
+
+* Bug Fix: microAmount not properly assigned in line item
+* Testing: Mock Ad Client and initial tests  
+
 0.1.1 (2020-11-24)
 ------------------